// frontend/src/api/models/Berufsgruppen.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bundestagswahlsystem API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { BerufsgruppenBerufsgruppenInner } from './BerufsgruppenBerufsgruppenInner';
import {
    BerufsgruppenBerufsgruppenInnerFromJSON,
    BerufsgruppenBerufsgruppenInnerFromJSONTyped,
    BerufsgruppenBerufsgruppenInnerToJSON,
    BerufsgruppenBerufsgruppenInnerToJSONTyped,
} from './BerufsgruppenBerufsgruppenInner';

/**
 * 
 * @export
 * @interface Berufsgruppen
 */
export interface Berufsgruppen {
    /**
     * 
     * @type {Array<BerufsgruppenBerufsgruppenInner>}
     * @memberof Berufsgruppen
     */
    berufsgruppen?: Array<BerufsgruppenBerufsgruppenInner>;
}

/**
 * Check if a given object implements the Berufsgruppen interface.
 */
export function instanceOfBerufsgruppen(value: object): value is Berufsgruppen {
    return true;
}

export function BerufsgruppenFromJSON(json: any): Berufsgruppen {
    return BerufsgruppenFromJSONTyped(json, false);
}

export function BerufsgruppenFromJSONTyped(json: any, ignoreDiscriminator: boolean): Berufsgruppen {
    if (json == null) {
        return json;
    }
    return {
        
        'berufsgruppen': json['berufsgruppen'] == null ? undefined : ((json['berufsgruppen'] as Array<any>).map(BerufsgruppenBerufsgruppenInnerFromJSON)),
    };
}

export function BerufsgruppenToJSON(json: any): Berufsgruppen {
    return BerufsgruppenToJSONTyped(json, false);
}

export function BerufsgruppenToJSONTyped(value?: Berufsgruppen | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'berufsgruppen': value['berufsgruppen'] == null ? undefined : ((value['berufsgruppen'] as Array<any>).map(BerufsgruppenBerufsgruppenInnerToJSON)),
    };
}