// frontend/src/api/models/Partei.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bundestagswahlsystem API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Partei
 */
export interface Partei {
    /**
     * 
     * @type {number}
     * @memberof Partei
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof Partei
     */
    shortname: string;
    /**
     * 
     * @type {string}
     * @memberof Partei
     */
    name?: string;
}

/**
 * Check if a given object implements the Partei interface.
 */
export function instanceOfPartei(value: object): value is Partei {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('shortname' in value) || value['shortname'] === undefined) return false;
    return true;
}

export function ParteiFromJSON(json: any): Partei {
    return ParteiFromJSONTyped(json, false);
}

export function ParteiFromJSONTyped(json: any, ignoreDiscriminator: boolean): Partei {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'shortname': json['shortname'],
        'name': json['name'] == null ? undefined : json['name'],
    };
}

export function ParteiToJSON(json: any): Partei {
    return ParteiToJSONTyped(json, false);
}

export function ParteiToJSONTyped(value?: Partei | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'shortname': value['shortname'],
        'name': value['name'],
    };
}